pyproject.toml
[tool.poetry]
name = "chat-negotiation"
version = "0.1.0"
description = ""
authors = ["gilbertofp16 <gilbertofp16@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10, <=3.13"
streamlit = "^1.32.2"
chromadb = "0.5.23"
pypdf = "^4.0.1"
pytesseract = "^0.3.10"
langfuse = "^2.15.0"
python-dotenv = "^1.0.1"
importlib-metadata = ">=4.13.0"
openai = "1.71.0"
litellm = "1.74.9"
langchain = "^0.3.27"
langchain-google-genai = "^2.1.9"
crewai = "^0.165.1"
langchain-community = "^0.3.27"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"

[tool.poetry.scripts]
ingest = "src.ingestion:ingest_book_pipeline"
trace-check = "src.observability:check_traces"
clean = "src.cleanup:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true